# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Naira Galaxy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
  fullName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  fullName: string;
  isAdmin: boolean;
  totalCoins: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface MiningStatus {
  date: string;
  minedButtons: number[];
  coinsEarnedToday: number;
  dailyLimit: number;
  canMineMore: boolean;
}

export interface CompleteMiningRequest {
  buttonIndex: number;
}

export interface MiningResult {
  coinsEarned: number;
  totalCoinsToday: number;
  totalCoins: number;
  minedButtons: number[];
  canMineMore: boolean;
}

export interface WalletBalance {
  totalCoins: number;
  pendingWithdrawal: number;
  availableBalance: number;
}

export interface WithdrawalRequest {
  amount: number;
  bankName: string;
  accountNumber: string;
  accountName: string;
}

export interface WithdrawalResponse {
  id: number;
  status: string;
  message: string;
}

export interface WithdrawalItem {
  id: number;
  amount: number;
  bankName: string;
  accountNumber: string;
  accountName: string;
  status: string;
  createdAt: string;
}

export interface AdminUser {
  id: number;
  username: string;
  email: string;
  fullName: string;
  totalCoins: number;
  createdAt: string;
  lastMiningDate?: string;
  todayCoins?: number;
}

export interface AdminWithdrawal {
  id: number;
  userId: number;
  username: string;
  fullName: string;
  amount: number;
  bankName: string;
  accountNumber: string;
  accountName: string;
  status: string;
  createdAt: string;
}

export type UpdateWithdrawalStatusRequestStatus =
  (typeof UpdateWithdrawalStatusRequestStatus)[keyof typeof UpdateWithdrawalStatusRequestStatus];

export const UpdateWithdrawalStatusRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  paid: "paid",
} as const;

export interface UpdateWithdrawalStatusRequest {
  status: UpdateWithdrawalStatusRequestStatus;
}