# lib/api-zod/src/generated/types/updateWithdrawalStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Naira Galaxy API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateWithdrawalStatusRequestStatus } from "./updateWithdrawalStatusRequestStatus";

export interface UpdateWithdrawalStatusRequest {
  status: UpdateWithdrawalStatusRequestStatus;
}