# lib/api-zod/src/generated/types/withdrawalItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Naira Galaxy API
 * OpenAPI spec version: 0.1.0
 */

export interface WithdrawalItem {
  id: number;
  amount: number;
  bankName: string;
  accountNumber: string;
  accountName: string;
  status: string;
  createdAt: string;
}