# lib/api-zod/src/generated/types/withdrawalResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Naira Galaxy API
 * OpenAPI spec version: 0.1.0
 */

export interface WithdrawalResponse {
  id: number;
  status: string;
  message: string;
}